api: migrate transaction/budget money types to Decimal; refresh SQLx cache

+++ jive-api/src/services/transaction_service.rs
@@ -1,5 +1,6 @@
 use crate::error::{ApiError, ApiResult};
 use crate::models::transaction::{Transaction, TransactionCreate, TransactionType};
+use rust_decimal::Decimal;
 use chrono::{DateTime, Utc};
 use sqlx::PgPool;
 use std::collections::HashMap;
@@ -28,7 +29,7 @@ impl TransactionService {
         let data_snapshot = data.clone();
 
         // 获取账户当前余额
-        let current_balance: Option<(f64,)> =
+        let current_balance: Option<(Decimal,)> =
             sqlx::query_as("SELECT current_balance FROM accounts WHERE id = $1 FOR UPDATE")
                 .bind(data.account_id)
                 .fetch_optional(&mut *tx)
@@ -127,7 +128,7 @@ impl TransactionService {
         target_account_id: Uuid,
     ) -> ApiResult<()> {
         // 获取目标账户余额
-        let target_balance: Option<(f64,)> =
+        let target_balance: Option<(Decimal,)> =
             sqlx::query_as("SELECT current_balance FROM accounts WHERE id = $1 FOR UPDATE")
                 .bind(target_account_id)
                 .fetch_optional(&mut **tx)
@@ -190,14 +191,14 @@ impl TransactionService {
             .map_err(|e| ApiError::DatabaseError(e.to_string()))?;
 
         let mut created_transactions = Vec::new();
-        let mut account_balances: HashMap<Uuid, f64> = HashMap::new();
+        let mut account_balances: HashMap<Uuid, Decimal> = HashMap::new();
 
         // 预加载所有相关账户的余额
         for trans in &transactions {
             if let std::collections::hash_map::Entry::Vacant(e) =
                 account_balances.entry(trans.account_id)
             {
-                let balance: Option<(f64,)> =
+                let balance: Option<(Decimal,)> =
                     sqlx::query_as("SELECT current_balance FROM accounts WHERE id = $1")
                         .bind(trans.account_id)
                         .fetch_optional(&mut *tx)
@@ -280,7 +281,7 @@ impl TransactionService {
     /// 智能分类交易
     pub async fn auto_categorize(&self, transaction_id: Uuid) -> ApiResult<Option<Uuid>> {
         // 获取交易信息
-        let transaction: Option<(String, Option<String>, f64)> =
+        let transaction: Option<(String, Option<String>, Decimal)> =
             sqlx::query_as("SELECT payee, notes, amount FROM transactions WHERE id = $1")
                 .bind(transaction_id)
                 .fetch_optional(&self.pool)
@@ -381,10 +382,10 @@ impl TransactionService {
 #[derive(Debug, sqlx::FromRow)]
 pub struct TransactionStatistics {
     pub total_count: i64,
-    pub total_income: Option<f64>,
-    pub total_expense: Option<f64>,
-    pub net_amount: Option<f64>,
-    pub avg_expense: Option<f64>,
-    pub max_expense: Option<f64>,
+    pub total_income: Option<Decimal>,
+    pub total_expense: Option<Decimal>,
+    pub net_amount: Option<Decimal>,
+    pub avg_expense: Option<Decimal>,
+    pub max_expense: Option<Decimal>,
     pub active_days: Option<i64>,
 }
